feat: adjust axis, clean up padding, make legend toggleable button
