+++ package.json
@@ -27,6 +27,7 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "d3": "^7.9.0",
+    "framer-motion": "^12.4.7",
     "fs": "^0.0.1-security",
     "lucide-react": "^0.439.0",
     "memfs": "^4.14.0",
@@ -35,6 +36,7 @@
     "polars": "^0.0.1",
     "react": "^18",
     "react-dom": "^18",
+    "react-icons": "^5.5.0",
     "react-plaid-link": "^3.6.1",
     "recharts": "^2.15.0",
     "tailwind-merge": "^2.5.2",
